% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alfred_series.R
\name{get_alfred_series}
\alias{get_alfred_series}
\title{This function estimates a multiplicative mixed-frequency GARCH model}
\usage{
get_alfred_series(series_id, series_name)
}
\arguments{
\item{series_id, }{series_name.}
}
\description{
This function estimates a multiplicative mixed-frequency GARCH model
}
\keyword{alfred}

