% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alfred_series.R
\name{get_fred_series}
\alias{get_fred_series}
\title{Accessing FRED}
\usage{
get_fred_series(series_id, series_name = NULL,
    observation_start = NULL, observation_end = NULL)
}
\arguments{
\item{series_id}{FRED times series ID.}

\item{series_name}{Choose a name for the series column in output. Default: series_id.}

\item{observation_start}{Date of first observation in "yyyy-mm-dd" format. Default: Earliest observation available.}

\item{observation_end}{Date of last observation in "yyyy-mm-dd" format. Default: Last observation available.}
}
\description{
This function can pull time series from the FRED database: \url{https://fred.stlouisfed.org}
}
\examples{
get_fred_series("INDPRO", "indpro")
}
\keyword{fred}
